pyproject.toml
[tool.poetry]
name = "apis-instance-nsvis"
version = "0.1.0"
description = ""
authors = ["Birger Schacht <birger.schacht@oeaw.ac.at>", "Matthias Schlögl <matthias.schloegl@oeaw.ac.at>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apis-core-rdf = "0.31.1"
apis-acdhch-default-settings = "1.5.0"
psycopg2 = "^2.9"
django-auditlog = "^3.0.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
